flytt OppgaveType til egen fil
